// src/app/shared/app/models/Production/i-edit-commissions.ts
export interface IEditCommissions
{
    sNo?: number
    clientNo?: number
    clientName?: string
    producer?: string
    insurComp?: string
    className?: string
    accNo?: string
    policyNo?: string
    compCommPerc?: string
    producerCommPerc?: string
    issueDate?: Date
    periodFrom?: Date
    periodTo?: Date,
    identity?: string
}